Creation of user, features, comment, vote tables

+++ src/main/java/com/freshvotes/domain/Features.java
@@ -6,38 +6,44 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
-public class Features {
-
-    private Integer id;
-    private String title;
-    private String description;
-    private String status;
-    
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Integer getId() {
-        return id;
-    }
-    public void setId(Integer id) {
-        this.id = id;
-    }
-    public String getTitle() {
-        return title;
-    }
-    public void setTitle(String title) {
-        this.title = title;
-    }
-    public String getDescription() {
-        return description;
-    }
-    public void setDescription(String description) {
-        this.description = description;
-    }
-    public String getStatus() {
-        return status;
-    }
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    
+public class Features
+{
+  private Long id;
+  private String title;
+  private String description;
+  private String status;
+   
+  @Id @GeneratedValue(strategy=GenerationType.IDENTITY)
+  public Long getId()
+  {
+    return id;
+  }
+  public void setId(Long id)
+  {
+    this.id = id;
+  }
+  public String getTitle()
+  {
+    return title;
+  }
+  public void setTitle(String title)
+  {
+    this.title = title;
+  }
+  public String getDescription()
+  {
+    return description;
+  }
+  public void setDescription(String description)
+  {
+    this.description = description;
+  }
+  public String getStatus()
+  {
+    return status;
+  }
+  public void setStatus(String status)
+  {
+    this.status = status;
+  }
 }
